DrawBounds work now.
For now it is fixed by making start to awake. This fixes the problem, because the problem was from the order of process in calculating positions and etc. By making it awake, the position and scale calculation is done without making the obj child of ModelManager, and completely do not need to care about the relationship scale, position problem.